Introduces a full-featured Tour Guide module:
- Adds domain entities (TourGuide, GuideTourPackage, GuideCity, GuideRequest, GuideOffer) and supporting enums.
- Implements CQRS commands/queries, validation, and error handling for tour guide and package management.
- Adds AutoMapper profiles and DTOs for API contracts.
- Creates TourGuideController with endpoints for CRUD operations.
- Registers new repositories and services in DI.
- Adds EF Core repositories and entity configurations.
- Updates ApplicationDbContext with new DbSets.
- Refactors namespaces (Presistances → Persistences), resolves merge conflicts, and improves code style.
- Updates GeoPoint to use decimal coordinates.
- Suppresses warning NU1903 in project files.

+++ Fayora.Domain/ValueObjects/GeoPoint.cs
@@ -4,10 +4,10 @@ namespace Fayora.Domain.ValueObjects;
 
 public class GeoPoint : ValueObject
 {
-    public double Latitude { get; private set; }
-    public double Longitude { get; private set; }
+    public decimal Latitude { get; private set; }
+    public decimal Longitude { get; private set; }
 
-    public GeoPoint(double latitude, double longitude)
+    public GeoPoint(decimal latitude, decimal longitude)
     {
         Latitude = latitude;
         Longitude = longitude;
